quick addMetadata.py fix

+++ addMetadata.py
@@ -11,7 +11,7 @@ import os
 try:
     filename = sys.argv[1]
     startval = float(sys.argv[2])
-    endval = int(sys.argv[3])
+    endval = float(sys.argv[3])
 
 except:
     print("This script assigns metadata of 1 inbetween the bounds and 0 elsewhere")
